+++ packages/say-ar-design-plugin/src/plugin/models/measure-concept.ts
@@ -1,0 +1,7 @@
+export default interface MeasureConcept {
+  id: string | null;
+  uri: string;
+  label: string;
+  templateString: string;
+  rawTemplateString: string;
+}