+++ packages/say-ar-design-plugin/src/plugin/models/variable.ts
@@ -1,0 +1,9 @@
+export default interface Variable {
+  id: string | null;
+  type: 'text' | 'number' | 'date' | 'codelist' | 'location';
+
+  label: string;
+  uri: string;
+  source: string;
+  codelist?: string;
+}